+++ src/main/java/racingcar/dao/GameLogDao.java
@@ -11,7 +11,7 @@ public class GameLogDao {
         this.jdbcTemplate = jdbcTemplate;
     }
 
-    public void insert(final int gameNumber,final String playerName,final int resultPosition) {
+    public void insert(final long gameNumber,final String playerName,final int resultPosition) {
         String sql = "insert into game_log (game_number, player_name, result_position) values (?, ?, ?)";
         jdbcTemplate.update(sql, gameNumber, playerName, resultPosition);
     }
